Added component for Examples
